feature: Add Coordinator and appBarLayout
Now the MainActivity layout (activity_article_list) contains a
Coordinator and appBarLayout. There is more work to do regarding the
scroll behavior.

Besides a missing file from the previous commit was added, in order to
ensure the correct fab button share functionality.

This commit modifies the layout file mentioned above and the
ArticleDetailFragment to share a custom text instead of a default one.