bug; Fix app performance
The application performance was terrible, taking sometimes up to 10
seconds to load an article and not responding to touch events.

The issue is related to the size of the article body, so as workaround
the body length as been limited to 1000 as suggested in a thread chat at
Udacity.

The commit modifies ArticleDetailFragment.java to implement the
workaround.